[runtime] Don't throw exceptions when checking if a token reference exists (and not finding any). Fixes #3830.
It's not necessarily bad to not be able to find a token reference for a class:
in particular not if we're just checking if a token reference exists. In this
case, don't throw any exceptions, so that the fall-back code path (if no token
references were found) can execute properly.

This scenario occurs when all the following are true:

* The runtime runs into a native object that is exposed in managed as a protocol.
* The native type's managed type does not implement the protocol.
* The dynamic registrar is being used.

Fixes https://github.com/xamarin/xamarin-macios/issues/3830.