[monotouch-test] Fix WeakReferenceTest.WeakTest to work on watchOS/LLVM.
Change WeakReferenceTest.WeakTest so that it doesn't fetch values that should
be garbage collected on the main thread.

Doing so on the main thread may cause those values to stay in registers as
temporary values, thus preventing the garbage collector from collecting them.

Instead do the fetching in a background thread, whose stack won't exist
anymore once it's finished.

Fixes this test failure when running on watchOS device in release mode:

    [FAIL] WeakReferenceTest.WeakTest : 't.Obj4' should be null

The reason it only happens on watchOS in release mode, is probably because
LLVM puts temporary values in different registers than Mono's AOT compiler
does.