[tests] Fix tests that depend on tool output
Lot of tests depend on comparing text in tool output. In xbuild, this
was logged with `MessageImportance.Normal` but in msbuild the default is
`MessageImportance.Low`. So, now the tool output doesn't show up in the
normal verbosity log, causing the build to fail.

Due to the number of tests depending on this, the default is changed to
`diagnosticMSBuild = true`.

A better way would probably (future) be to use a custom logger and get the output
from the logging events.