Improve IsCustomType performance
While NSObject may now flag internally whether it is a custom type, this
still has to be looked up each time a new instance of that type is created.

The dynamic registrar has a fast lookup for whether a type is a custom type
or not, but IsCustomType first checks the static registrar regardless.
GetClassHandle already caches the class handles, so this extends it to include
whether or not the type is custom so then IsCustomType can use the cached path.