[xharness] Handle failure to create a device pair for the simulator better. Fixes xamarin/maccore#830.
Detect if we failed to create a simulator device pair due to trying with a
watch that's already paired. We were already detecting this scenario if the
watch was a member of an available pair, but not if the watch was a member of
an _unavailable_ pair (since mlaunch doesn't list unavailable device pairs).

So detect this scenario from the error message simctl gives us instead, and in
that case create a new watch device and try to create the pair again.

Fixes https://github.com/xamarin/maccore/issues/830.