[xtro] Fix "missing unclassified" issue
- On 10.14 we were seeing unclassified issues appear that were not
on 10.13 builds.
- It turns out each instance involved a framework with a different
capitalization (PDF -> Pdf, MIDI -> Midi, WLAN -> WLan) in our bindings
than the header.
- For some reason that still isn't 100% clear, we'd see both our binding
and some clang item and register by capitalizations in the list in Log.cs
- Then we'd process the first list, write it down to disk, process the
second list (which is empty) by deleting that file that we just wrote down
- In any case, we can just teach On () to ignore capitalization when looking
for a list to prevent this madness.