[macos][quartzcomposer] Update for beta 4
Apple re-added some constants that were dropped without any clue (and we
filed a rdar for it since it's a breaking change to remove them).

It's now clear that it's deprecated and why
`QC_GL_DEPRECATED(10_5, 10_14);`

so we can update the attribute (and fix the version).

note: also fix extra `;` in `CBUUIDValidRangeString` tests while removing
the special case added earlier (from beta 1)