[src] Use generated rsp files when executing the generator.
This is step 1 in not having to generate generator.csproj, since we can
reference the rsp files statically instead of injecting the actual command
line arguments to the generator for the debugging run configurations.

It also makes sure the project's run configurations are always up to date,
since they're using the exact same input as the command line build.