[tests] Ignore RuntimeTest.MainThreadDeallocationTest if all optimizations are enabled and not linking everything. Fixes #4566.
This test contains code that requires the dynamic registrar unless the code is
optimized, and it's only optimized if the assembly is linked, which only
happens if linking all assemblies.

So disable the test if dynamic registrar is removed (which happens when all
optimizations are enabled) and the assembly isn't linked.

Fixes https://github.com/xamarin/xamarin-macios/issues/4566.