Add a tool that will allow to generate a project per set of assemblies.
The tool provides all the different steps to generate the project file
as well as to a small registration code class that will ensure that the
assembly is not linked away.

The make file now create a project per definition (in the top of the
makefile) and compiles them. Next, get xharness to run each of the
generated apps.